+++ mercator-aws/src/main/java/org/lendingclub/mercator/aws/AvailabilityZoneScanner.java
@@ -44,7 +44,7 @@ public class AvailabilityZoneScanner extends AbstractEC2Scanner{
 	}
 
 	@Override
-	public void doScan() {
+	protected void doScan() {
 		
 
 		GraphNodeGarbageCollector gc = newGarbageCollector().region(getRegion()).label("AwsAvailabilityZone");